chore: remove old linters command
