+++ pkg/commands/run.go
@@ -524,3 +524,16 @@ func initConfigFileFlagSet(fs *pflag.FlagSet, cfg *config.Run) {
 	fs.StringVarP(&cfg.Config, "config", "c", "", wh("Read config from file path `PATH`"))
 	fs.BoolVar(&cfg.NoConfig, "no-config", false, wh("Don't read config file"))
 }
+
+// --- Related to linters but use here.
+
+func initLintersFlagSet(fs *pflag.FlagSet, cfg *config.Linters) {
+	fs.StringSliceVarP(&cfg.Disable, "disable", "D", nil, wh("Disable specific linter"))
+	fs.BoolVar(&cfg.DisableAll, "disable-all", false, wh("Disable all linters"))
+	fs.StringSliceVarP(&cfg.Enable, "enable", "E", nil, wh("Enable specific linter"))
+	fs.BoolVar(&cfg.EnableAll, "enable-all", false, wh("Enable all linters"))
+	fs.BoolVar(&cfg.Fast, "fast", false, wh("Enable only fast linters from enabled linters set (first run won't be fast)"))
+	fs.StringSliceVarP(&cfg.Presets, "presets", "p", nil,
+		wh(fmt.Sprintf("Enable presets (%s) of linters. Run 'golangci-lint help linters' to see "+
+			"them. This option implies option --disable-all", strings.Join(lintersdb.AllPresets(), "|"))))
+}
